backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.18
pdfplumber==0.11.4
Pillow==11.0.0
pdf2image==1.17.0
httpx==0.27.2
aiosqlite==0.20.0
pydantic-settings==2.6.1
